chore: rename files to follow standards
